requirements.txt
pyreadline
requests
beautifulsoup4
html5lib
docopt
bs4
python-nmap
